refactor: rename tests/xxx_tests.rs files to tests/xxx.rs
