+++ libwayshot/src/tests/error.rs
