+++ libwayshot/src/tests/image_util.rs
