+++ libwayshot/src/tests/mod.rs
@@ -1,18 +1,18 @@
 //! Private test module. All unit tests for libwayshot live here to keep main source files focused.
 
-mod convert_tests;
-mod error_tests;
-mod image_util_tests;
-mod screencopy_tests;
+mod convert;
+mod error;
+mod image_util;
+mod screencopy;
 
 #[cfg(all(test, unix))]
-mod output_tests;
+mod output;
 
 #[cfg(all(test, unix))]
-mod region_tests;
+mod region;
 
 #[cfg(all(test, unix))]
-mod dispatch_tests;
+mod dispatch;
 
 #[cfg(all(test, unix))]
-mod lib_tests;
+mod lib;
