+++ libwayshot/src/tests/output.rs
