+++ libwayshot/src/tests/screencopy.rs
